backend/pyproject.toml
[tool.poetry]
name = "cloud_api"
version = "0.1.0"
description = ""
authors = ["ArturZhukovets <bnlvdn7483@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
requests = "^2.31.0"
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "^0.29.0"
uvicorn = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
